requirements.txt
fastapi
uvicorn[standard]
tenacity
google-genai
pydantic>=2.5.0
pydantic-settings
pytest
pytest-asyncio
pytest-mock
httpx
python-dotenv
